requirements.txt
flask
werkzeug
torch
torchvision
scikit-learn
Pillow
numpy
sqlalchemy
vercel-wsgi